requirements.txt
absl-py==2.1.0
acnportal==0.3.3
aiosignal==1.3.2
ale-py==0.10.1
astunparse==1.6.3
attrs==24.3.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
box2d-py==2.3.5
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chex==0.1.88
clarabel==0.9.0
click==8.1.7
cloudpickle==1.2.2
contourpy==1.3.0
cramjam==2.9.1
cvxpy==1.6.0
cycler==0.12.1
Cython==0.29.37
decorator==5.1.1
dill==0.3.9
distlib==0.3.8
dm-tree==0.1.8
docker-pycreds==0.4.0
-e git+https://github.com/google/dopamine@bec5f4e108b0572e58fc1af73136e978237c8463#egg=dopamine_rl
etils==1.5.2
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
flatbuffers==24.12.23
flax==0.8.5
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.8.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.68.1
gym==0.15.7
gym-notices==0.0.8
gymnasium==1.0.0
gymnasium-robotics==1.3.1
h5py==3.12.1
humanize==4.11.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ivon-opt==0.1.3
jax==0.4.30
jax-jumpy==1.0.0
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpi4py==4.0.1
mpmath==1.3.0
msgpack==1.1.0
mujoco==2.2.0
mujoco-py==2.1.2.14
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.6.4
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pettingzoo==1.24.3
pillow==11.0.0
pipenv==2024.0.1
pipfile==0.0.2
platformdirs==4.3.2
proglog==0.1.10
progressbar2==4.5.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.1.0
pyglet==1.5.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-snappy==0.7.3
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdldl==0.1.7.post4
ray==2.40.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.13.1
scs==3.2.7
sentry-sdk==2.17.0
setproctitle==1.3.3
shellingham==1.5.4
Shimmy==0.2.1
six==1.17.0
smmap==5.0.1
stable_baselines3==2.4.0
sustaingym==0.1.7
swig==4.2.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
tensorstore==0.1.69
termcolor==2.5.0
tf-slim==1.1.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.8.30
toml==0.10.2
toolz==1.0.0
torch==2.5.1
tqdm==4.66.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.26.4
wandb==0.18.5
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
zmq==0.0.0
